Extract API model helpers
