+++ js/api.js
@@ -1,7 +1,6 @@
 // @ts-check
-// api.js — AI provider management, API calls (OpenRouter, Venice, Routstr, PPQ, Local, Custom)
+// api.js - AI provider management, API calls (OpenRouter, Venice, Routstr, PPQ, Local, Custom)
 
-import { getModelPricing } from './schema.js';
 import { isDebugMode } from './utils.js';
 import {
   FETCH_REQUEST_TIMEOUT_MS,
@@ -28,7 +27,6 @@ import {
   getVeniceModelDisplay,
   getVeniceE2EE,
   setVeniceE2EE,
-  modelSupportsVeniceE2EE,
   readStoredArray,
   syncVeniceModelSelection,
   veniceModelsCacheStale,
@@ -64,6 +62,21 @@ import {
   setCustomApiModel,
   getCustomApiModelDisplay,
 } from './api-provider-storage.js';
+import {
+  deduplicateModels,
+  fetchOpenRouterModelPricing,
+  fetchOpenRouterModels,
+  fetchVeniceModels,
+  getActiveModelDisplay,
+  getActiveModelId,
+  isRecommendedModel,
+  needsMaxCompletionTokens,
+  renderModelPricingHint,
+  supportsVision,
+  supportsWebSearch,
+  validateOpenRouterKey,
+  validateVeniceKey,
+} from './api-models.js';
 
 /** @typedef {Window & typeof globalThis & {
  *   _veniceE2EE?: any,
@@ -79,6 +92,21 @@ export {
   FETCH_REQUEST_TIMEOUT_MS,
   STREAM_STALL_TIMEOUT_MS,
 } from './api-transport.js';
+export {
+  deduplicateModels,
+  fetchOpenRouterModelPricing,
+  fetchOpenRouterModels,
+  fetchVeniceModels,
+  getActiveModelDisplay,
+  getActiveModelId,
+  isRecommendedModel,
+  needsMaxCompletionTokens,
+  renderModelPricingHint,
+  supportsVision,
+  supportsWebSearch,
+  validateOpenRouterKey,
+  validateVeniceKey,
+} from './api-models.js';
 export {
   getAIProvider,
   setAIProvider,
@@ -140,16 +168,6 @@ function isTokenLimitFinish(reason) {
     || r === 'max_completion_tokens'
     || r.includes('token_limit')
     || r.includes('max token');
-}
-
-export function deduplicateModels(models, familyFn) {
-  const seen = {};
-  return models.filter(function(m) {
-    const fam = familyFn(m.id);
-    if (seen[fam]) return false;
-    seen[fam] = true;
-    return true;
-  });
 }
 
 const OPENROUTER_OAUTH_PREVIOUS_PROVIDER_KEY = 'or_previous_ai_provider';
@@ -283,213 +301,11 @@ export async function exchangeOpenRouterCode(code, returnedState) {
   sessionStorage.removeItem('or_oauth_state');
   return data.key;
 }
-// Curated: latest-gen medically capable models only (prefixes matched against IDs)
-const OPENROUTER_CURATED = [
-  'anthropic/claude-sonnet-4', 'anthropic/claude-opus-4',
-  'openai/gpt-5',
-  'google/gemini-3', 'google/gemini-2',
-  'deepseek/deepseek',
-  'qwen/qwen', 'qwen/qwq',
-  'x-ai/grok',
-];
-// ─── Recommended models for medical analysis ───
-// Update when a new generation launches. Each provider uses different ID formats:
-//   OpenRouter: "provider/model-version"  (dots: 4.6)
-//   Anthropic:  "claude-model-version"    (hyphens: 4-6, with date suffix)
-//   Venice:     "model-version"           (hyphens: 4-6, no provider prefix)
-// To check current IDs, run in console:
-//   JSON.parse(localStorage.getItem('labcharts-openrouter-models')||'[]').map(m=>m.id)
-const OPENROUTER_RECOMMENDED = [
-  'anthropic/claude-sonnet-4.6', 'anthropic/claude-opus-4.7',
-  'openai/gpt-5.5', 'openai/gpt-5.4',
-  'google/gemini-3.1-pro',
-  'x-ai/grok-4',
-];
-// Routstr uses bare model IDs (no provider prefix, dots: claude-sonnet-4.6)
 const ROUTSTR_CURATED = ['claude-', 'gpt-5', 'gpt-4', 'gemini-3', 'gemini-2', 'grok-4', 'grok-3', 'llama-', 'qwen', 'deepseek-', 'mistral-', 'mimo-'];
 const ROUTSTR_RECOMMENDED = ['claude-sonnet-4.6', 'claude-opus-4.7', 'gpt-5.5', 'gpt-5.4', 'gemini-3.1-pro', 'grok-4'];
-// PPQ uses bare model IDs (same as Routstr)
-// private/ models (Tinfoil TEE) listed in API but require EHBP protocol, not standard completions
 const PPQ_CURATED = ['claude-', 'gpt-5', 'gpt-4', 'gpt-oss', 'gemini-3', 'gemini-2', 'grok-', 'llama-', 'qwen', 'deepseek-', 'mistral-', 'kimi', 'perplexity'];
 const PPQ_RECOMMENDED = ['claude-sonnet-4.6', 'claude-opus-4.7', 'gpt-5.5', 'gpt-5.4', 'gemini-3-flash-preview', 'grok-4'];
 const PPQ_EXCLUDE = ['codex', 'audio', 'image', 'embed', 'tts', 'whisper', 'video', 'nano-banana'];
-export function isRecommendedModel(provider, modelId) {
-  if (provider === 'openrouter') return OPENROUTER_RECOMMENDED.some(function(prefix) { return modelId.startsWith(prefix); });
-  if (provider === 'venice') {
-    if (modelId.startsWith('e2ee-')) return /qwen3-5-122b|gpt-oss-120b|qwen3-30b|glm-5/.test(modelId);
-    // claude-(sonnet-4-6|opus-4-7) is intentionally narrow — Anthropic ships
-    // each model as its own version. When sonnet-4-7 / opus-4-8 land, broaden
-    // the alternation rather than back to (sonnet|opus)-4-X (would over-match
-    // older versions). gpt-5[2345] tracks 5.2/5.3/5.4/5.5 — extend digits as
-    // OpenAI ships new minor versions.
-    return /^(claude-(sonnet-4-6|opus-4-7)|openai-gpt-5[2345](-codex)?|gemini-3(-1)?-pro|grok-4[1-9]?)(-|$)/.test(modelId);
-  }
-  if (provider === 'routstr') return ROUTSTR_RECOMMENDED.some(function(r) { return modelId === r || modelId.startsWith(r); });
-  if (provider === 'ppq') return PPQ_RECOMMENDED.some(function(r) { return modelId === r || modelId.startsWith(r); });
-  return false; // Ollama — local models, can't tier
-}
-export function getActiveModelId(provider = getAIProvider()) {
-  if (provider === 'venice') return getVeniceModel();
-  if (provider === 'openrouter') return getOpenRouterModel();
-  if (provider === 'routstr') return getRoutstrModel();
-  if (provider === 'ppq') return getPpqModel();
-  if (provider === 'custom') return getCustomApiModel();
-  return getOllamaMainModel();
-}
-export function getActiveModelDisplay(provider = getAIProvider()) {
-  if (provider === 'venice') return getVeniceModelDisplay();
-  if (provider === 'openrouter') return getOpenRouterModelDisplay();
-  if (provider === 'routstr') return getRoutstrModelDisplay();
-  if (provider === 'ppq') return getPpqModelDisplay();
-  if (provider === 'custom') return getCustomApiModelDisplay();
-  return getOllamaMainModel();
-}
-// Exclude specialized variants not suited for medical analysis
-const OPENROUTER_EXCLUDE = ['codex', 'audio', 'image', 'oss', 'safeguard', 'coder'];
-export async function fetchOpenRouterModels(key) {
-  try {
-    const res = await fetch('https://openrouter.ai/api/v1/models', {
-      headers: { 'Authorization': 'Bearer ' + (key || getOpenRouterKey()) }
-    });
-    if (!res.ok) return [];
-    const json = await res.json();
-    // Filter to curated medically capable models, exclude specialized variants
-    const all = (json.data || []).filter(function(m) {
-      if (!m.id) return false;
-      if (OPENROUTER_EXCLUDE.some(function(ex) { return m.id.includes(ex); })) return false;
-      return OPENROUTER_CURATED.some(function(prefix) { return m.id.startsWith(prefix); });
-    }).sort(function(a, b) { return (a.name || a.id).localeCompare(b.name || b.id); });
-    // Deduplicate: strip date/size suffixes after provider/ prefix
-    const models = deduplicateModels(all, function(id) {
-      return id.replace(/:\d{4}-\d{2}-\d{2}$/, '').replace(/-\d{8}$/, '');
-    });
-    // Sort recommended models first, then alphabetical within each group
-    models.sort(function(a, b) {
-      const aRec = OPENROUTER_RECOMMENDED.some(function(p) { return a.id.startsWith(p); });
-      const bRec = OPENROUTER_RECOMMENDED.some(function(p) { return b.id.startsWith(p); });
-      if (aRec !== bRec) return aRec ? -1 : 1;
-      return (a.name || a.id).localeCompare(b.name || b.id);
-    });
-    // Extract per-million-token pricing from API response
-    const pricingCache = {};
-    for (const m of models) {
-      if (m.pricing && m.pricing.prompt && m.pricing.completion) {
-        pricingCache[m.id] = {
-          input: parseFloat(m.pricing.prompt) * 1_000_000,
-          output: parseFloat(m.pricing.completion) * 1_000_000
-        };
-      }
-    }
-    localStorage.setItem('labcharts-openrouter-pricing', JSON.stringify(pricingCache));
-    // Cache vision-capable model IDs (architecture.modality contains "image->text" or similar)
-    const visionIds = (json.data || []).filter(function(m) {
-      if (!m.id || !m.architecture) return false;
-      const modality = m.architecture.modality || '';
-      return modality.includes('image');
-    }).map(function(m) { return m.id; });
-    localStorage.setItem('labcharts-openrouter-vision-models', JSON.stringify(visionIds));
-    localStorage.setItem('labcharts-openrouter-models', JSON.stringify(models));
-    if (!localStorage.getItem('labcharts-openrouter-model') && models.length) {
-      const claude = models.find(function(m) { return m.id === 'anthropic/claude-sonnet-4.6'; });
-      if (claude) setOpenRouterModel(claude.id);
-    }
-    return models;
-  } catch (e) { return []; }
-}
-/** Fetch and cache pricing for a custom OpenRouter model not in the curated list */
-export async function fetchOpenRouterModelPricing(modelId) {
-  if (!modelId) return null;
-  const existing = getOpenRouterPricing(modelId);
-  if (existing) return existing;
-  try {
-    const res = await fetch('https://openrouter.ai/api/v1/models', {
-      headers: { 'Authorization': 'Bearer ' + getOpenRouterKey() }
-    });
-    if (!res.ok) return null;
-    const json = await res.json();
-    // Exact match first, then fuzzy (dots vs dashes, date suffixes)
-    const norm = s => s.replace(/\./g, '-').replace(/-\d{8}$/, '');
-    const model = (json.data || []).find(m => m.id === modelId)
-      || (json.data || []).find(m => norm(m.id) === norm(modelId));
-    if (!model?.pricing) return null;
-    const pricing = {
-      input: parseFloat(model.pricing.prompt || '0') * 1_000_000,
-      output: parseFloat(model.pricing.completion || '0') * 1_000_000
-    };
-    const cached = JSON.parse(localStorage.getItem('labcharts-openrouter-pricing') || '{}');
-    // Cache under both the API ID and the user-typed ID
-    cached[model.id] = pricing;
-    cached[modelId] = pricing;
-    localStorage.setItem('labcharts-openrouter-pricing', JSON.stringify(cached));
-    return pricing;
-  } catch (e) { /* fail silently */ }
-  return null;
-}
-
-export async function validateOpenRouterKey(key) {
-  try {
-    const res = await fetch('https://openrouter.ai/api/v1/models', {
-      headers: { 'Authorization': 'Bearer ' + key }
-    });
-    if (res.ok) return { valid: true };
-    if (res.status === 401) return { valid: false, error: 'Invalid API key' };
-    if (res.status === 429) return { valid: true };
-    const errBody = await res.json().catch(() => null);
-    const errMsg = errBody?.error?.message || `status ${res.status}`;
-    return { valid: false, error: `API error: ${errMsg}` };
-  } catch (e) {
-    return { valid: false, error: 'Cannot reach OpenRouter API: ' + e.message };
-  }
-}
-
-export function renderModelPricingHint(provider, modelId) {
-  if (provider === 'ollama') return '<span style="font-size:11px;color:var(--green)">Free (local)</span>';
-  if (provider === 'custom') return '';
-  const p = getModelPricing(provider, modelId);
-  if (p.input === 0 && p.output === 0) return '<span style="font-size:11px;color:var(--green)">Free</span>';
-  const pre = p.approx ? '~' : '';
-  return `<span style="font-size:11px;color:var(--text-muted)">${pre}$${p.input.toFixed(2)}/M in \u00b7 ${pre}$${p.output.toFixed(2)}/M out</span>`;
-}
-export async function fetchVeniceModels(key) {
-  try {
-    const res = await fetch('https://api.venice.ai/api/v1/models', {
-      headers: { 'Authorization': 'Bearer ' + (key || getVeniceKey()) }
-    });
-    if (!res.ok) return [];
-    const json = await res.json();
-    // Sort descending so latest version comes first per family
-    const allText = (json.data || []).filter(function(m) { return m.id && m.type === 'text'; }).sort(function(a, b) { return b.id.localeCompare(a.id); });
-    // Cache E2EE models separately. The capability flag is authoritative;
-    // keep a prefix fallback for older Venice responses that did not include it.
-    const e2eeList = allText.filter(modelSupportsVeniceE2EE);
-    localStorage.setItem('labcharts-venice-e2ee-models', JSON.stringify(e2eeList));
-    const e2eeIds = new Set(e2eeList.map(function(m) { return m.id; }));
-    const all = allText.filter(function(m) { return !e2eeIds.has(m.id) && !m.id.startsWith('e2ee-'); });
-    // Deduplicate: Venice curates Claude models (no date-stamped variants), so keep all.
-    // For others, strip size/date suffixes to collapse duplicates.
-    const models = deduplicateModels(all, function(id) {
-      if (id.startsWith('claude-')) return id;
-      return id.replace(/-\d{8}$/, '').replace(/-\d+[bB]$/, '');
-    });
-    // Re-sort alphabetically by display name
-    models.sort(function(a, b) { return (a.name || a.id).localeCompare(b.name || b.id); });
-    // Extract per-million-token pricing from model_spec
-    const pricingCache = {};
-    for (const m of allText) {
-      const p = m.model_spec && m.model_spec.pricing;
-      if (p && p.input && p.output) {
-        pricingCache[m.id] = { input: parseFloat(p.input.usd || 0), output: parseFloat(p.output.usd || 0) };
-      }
-    }
-    localStorage.setItem('labcharts-venice-pricing', JSON.stringify(pricingCache));
-    const visionIds = allText.filter(m => m.model_spec?.capabilities?.supportsVision).map(m => m.id);
-    localStorage.setItem('labcharts-venice-vision-models', JSON.stringify(visionIds));
-    localStorage.setItem('labcharts-venice-models', JSON.stringify(models));
-    localStorage.setItem('labcharts-venice-models-fetched-at', String(Date.now()));
-    syncVeniceModelSelection(models, e2eeList);
-    return models;
-  } catch (e) { return []; }
-}
 
 // ─── Proxy support ───
 // Only Custom API needs the proxy (arbitrary endpoints may lack CORS headers).
@@ -512,57 +328,6 @@ async function _fetchWithRetry(url, options, retries = 2, useProxy = true, reque
     directFetch: fetch,
     debug: isDebugMode,
   });
-}
-
-// ═══════════════════════════════════════════════
-// WEB SEARCH SUPPORT
-// ═══════════════════════════════════════════════
-export function supportsWebSearch(provider = getAIProvider()) {
-  if (provider === 'venice') return !isVeniceE2EEActive();
-  if (provider === 'routstr') return false;
-  if (provider === 'ppq') return true;
-  if (provider === 'custom') return false;
-  return provider === 'openrouter';
-}
-
-// ═══════════════════════════════════════════════
-// VISION SUPPORT
-// ═══════════════════════════════════════════════
-export function supportsVision() {
-  const provider = getAIProvider();
-  if (provider === 'openrouter') {
-    const modelId = getOpenRouterModel();
-    try {
-      const visionIds = JSON.parse(localStorage.getItem('labcharts-openrouter-vision-models') || '[]');
-      // Check exact match or prefix match (model IDs may have date suffixes)
-      return visionIds.some(function(vid) { return modelId === vid || modelId.startsWith(vid.replace(/:\d{4}-\d{2}-\d{2}$/, '')); });
-    } catch { return false; }
-  }
-  if (provider === 'venice') {
-    if (isVeniceE2EEActive()) return false;
-    const modelId = getVeniceModel();
-    try {
-      const visionIds = JSON.parse(localStorage.getItem('labcharts-venice-vision-models') || '[]');
-      return visionIds.some(function(vid) { return modelId === vid || modelId.startsWith(vid.replace(/-\d{8}$/, '')); });
-    } catch { return false; }
-  }
-  if (provider === 'routstr') {
-    const modelId = getRoutstrModel();
-    try {
-      const visionIds = JSON.parse(localStorage.getItem('labcharts-routstr-vision-models') || '[]');
-      return visionIds.some(function(vid) { return modelId === vid || modelId.startsWith(vid.replace(/-\d{8}$/, '')); });
-    } catch { return false; }
-  }
-  if (provider === 'ppq') {
-    const modelId = getPpqModel();
-    try {
-      const visionIds = JSON.parse(localStorage.getItem('labcharts-ppq-vision-models') || '[]');
-      return visionIds.some(function(vid) { return modelId === vid || modelId.startsWith(vid.replace(/-\d{8}$/, '')); });
-    } catch { return false; }
-  }
-  // Custom API / Local AI — optimistic (user's responsibility)
-  if (provider === 'custom') return true;
-  return true;
 }
 
 export async function callOllamaChat({ system, messages, maxTokens, onStream, signal }) {
@@ -658,18 +423,6 @@ export async function callOllamaChat({ system, messages, maxTokens, onStream, si
     const data = await res.json();
     return { text: data.message?.content || '', usage: { inputTokens: data.prompt_eval_count || 0, outputTokens: data.eval_count || 0 } };
   }
-}
-
-// ═══════════════════════════════════════════════
-// SHARED OPENAI-COMPATIBLE API HELPER
-// ═══════════════════════════════════════════════
-// GPT-5 family + o-series reasoning models reject `max_tokens` and require `max_completion_tokens`.
-// Matches bare ids (gpt-5.4, o1-mini) and provider-prefixed (openai/gpt-5, openai/o3).
-export function needsMaxCompletionTokens(modelId) {
-  if (!modelId) return false;
-  const id = String(modelId).toLowerCase();
-  const bare = id.includes('/') ? id.split('/').pop() : id;
-  return /^(gpt-5|o[1-9])([-.]|$)/.test(bare);
 }
 
 async function callOpenAICompatibleAPI(endpoint, key, model, providerName, { system, messages, maxTokens, onStream, signal, requestTimeoutMs, jsonMode }, extraHeaders = {}, { useProxy = true, extraBody = {} } = {}) {
@@ -963,22 +716,6 @@ export async function callOpenRouterAPI(opts) {
     { 'HTTP-Referer': window.location.origin, 'X-Title': 'getbased' },
     { extraBody }
   );
-}
-
-export async function validateVeniceKey(key) {
-  try {
-    const res = await fetch('https://api.venice.ai/api/v1/models', {
-      headers: { 'Authorization': 'Bearer ' + key }
-    });
-    if (res.ok) return { valid: true };
-    if (res.status === 401) return { valid: false, error: 'Invalid API key' };
-    if (res.status === 429) return { valid: true };
-    const errBody = await res.json().catch(() => null);
-    const errMsg = errBody?.error?.message || `status ${res.status}`;
-    return { valid: false, error: `API error: ${errMsg}` };
-  } catch (e) {
-    return { valid: false, error: 'Cannot reach Venice API: ' + e.message };
-  }
 }
 
 // ═══════════════════════════════════════════════
