+++ tests/test-image-utils.js
@@ -121,8 +121,9 @@ assert('parseLabPDFWithAIImages exported', typeof window.parseLabPDFWithAIImages
 console.log('9. Source Code');
 
 const apiSrc = await fetchWithRetry('js/api.js');
-assert('supportsVision function in api.js', apiSrc.includes('export function supportsVision'));
-assert('Vision models cached in fetchOpenRouterModels', apiSrc.includes('labcharts-openrouter-vision-models'));
+const apiModelsSrc = await fetchWithRetry('js/api-models.js');
+assert('supportsVision function in api-models.js', apiModelsSrc.includes('export function supportsVision'));
+assert('Vision models cached in fetchOpenRouterModels', apiModelsSrc.includes('labcharts-openrouter-vision-models'));
 assert('Ollama image normalization', apiSrc.includes('ollamaMsg.images = images'));
 
 const chatRenderSrc = await fetchWithRetry('js/chat-render.js');
